requirements.txt
Pillow==10.2.0
openpyxl==3.0.10